Size measurements for Setaria viridis

This dataset contains measurements of various parts and attributes for each branch of Setaria viridis. The variables are described as follows:

Usage

setaria_viridis

Format

A data frame with 54 rows and 5 variables:

root_number

A factor indicating which branches share the same root.

culm_length

Numeric. Length of the culm above ground, measured in centimeters (cm).

panicle_width

Numeric. Width of the panicle, measured in centimeters (cm).

panicle_length

Numeric. Length of the panicle, measured in centimeters (cm).

leaf_number

Integer. Number of leaves on the side branch.