setariaviridis: Field-Collected Data of Green Foxtail (Setaria vilidis)

Setaria viridis (green foxtail) is a common weed. This package contains measurements from individual branches of a wild Setaria viridis plant collected near the author's home. The data is intended for use in data analysis practice.

See also

Author

Maintainer: Keisuke ANDO ando@maslab.aitech.ac.jp (ORCID)